adapted to the jboss-op annotation
