+++ wildfly-config-gen/src/main/java/org/jboss/provisioning/plugin/wildfly/configgen/WfProvisionedConfigHandler.java
@@ -116,7 +116,7 @@ public class WfProvisionedConfigHandler implements ProvisionedConfigHandler {
                         if (!inAddr) {
                             if (paramFilter.accepts(paramName, j)) {
                                 final ManagedOp mop = new ManagedOp();
-                                mop.name = annotation.getName();
+                                mop.name = WfConstants.WRITE_ATTRIBUTE;
                                 mop.op = WRITE_ATTR;
                                 mop.addrPref = annotation.getElement(WfConstants.ADDR_PREF);
                                 mop.addrParams = addrParams;
@@ -141,7 +141,7 @@ public class WfProvisionedConfigHandler implements ProvisionedConfigHandler {
                 for (int i = 0; i < params.size(); i++) {
                     if (i % 2 == 0) {
                         final ManagedOp mop = new ManagedOp();
-                        mop.name = annotation.getName();
+                        mop.name = WfConstants.WRITE_ATTRIBUTE;
                         mop.op = WRITE_ATTR;
                         mop.addrPref = annotation.getElement(WfConstants.ADDR_PREF);
                         mop.addrParams = addrParams;
@@ -160,17 +160,13 @@ public class WfProvisionedConfigHandler implements ProvisionedConfigHandler {
     }
 
     private List<ManagedOp> createAddListManagedOperation(ResolvedFeatureSpec spec, FeatureAnnotation annotation) throws ProvisioningException {
-        return createManagedOperation(spec, annotation, LIST_ADD);
+        return createManagedOperation(spec, annotation, WfConstants.LIST_ADD, LIST_ADD);
     }
 
-    private List<ManagedOp> createAddManagedOperation(ResolvedFeatureSpec spec, FeatureAnnotation annotation) throws ProvisioningException {
-        return createManagedOperation(spec, annotation, OP);
-    }
-
-    private List<ManagedOp> createManagedOperation(ResolvedFeatureSpec spec, FeatureAnnotation annotation, int operation) throws ProvisioningException {
+    private List<ManagedOp> createManagedOperation(ResolvedFeatureSpec spec, FeatureAnnotation annotation, String name, int operation) throws ProvisioningException {
         final ManagedOp mop = new ManagedOp();
         mop.reset();
-        mop.name = annotation.getName();
+        mop.name = name;
         mop.op = operation;
         mop.addrPref = annotation.getElement(WfConstants.ADDR_PREF);
 
@@ -462,7 +458,9 @@ public class WfProvisionedConfigHandler implements ProvisionedConfigHandler {
 
         final List<FeatureAnnotation> annotations = spec.getAnnotations();
         for (FeatureAnnotation annotation : annotations) {
-            ops.addAll(nextAnnotation(spec, annotation));
+            if(annotation.getName().equals(WfConstants.JBOSS_OP)) {
+                ops.addAll(nextAnnotation(spec, annotation));
+            }
         }
     }
 
@@ -473,15 +471,14 @@ public class WfProvisionedConfigHandler implements ProvisionedConfigHandler {
             mop.line = annotation.getElement(WfConstants.LINE);
             return Collections.singletonList(mop);
         }
-        switch (annotation.getName()) {
-            case WfConstants.ADD:
-                return createAddManagedOperation(spec, annotation);
+        final String name = annotation.getElement(WfConstants.NAME);
+        switch (name) {
             case WfConstants.WRITE_ATTRIBUTE:
                 return createWriteAttributeManagedOperation(spec, annotation);
-                case WfConstants.LIST_ADD:
+            case WfConstants.LIST_ADD:
                 return createAddListManagedOperation(spec, annotation);
             default:
-                return createManagedOperation(spec, annotation, OP);
+                return createManagedOperation(spec, annotation, name, OP);
         }
     }
